chore: check proxy in refund widget
