+++ apps/cowswap-frontend/src/modules/cowShed/containers/RecoverFundsWidget/index.tsx
@@ -2,12 +2,8 @@ import { atom, useAtom } from 'jotai/index'
 import { ReactNode, useCallback } from 'react'
 
 import { TokenWithLogo } from '@cowprotocol/common-const'
-import { getCurrencyAddress, getIsNativeToken } from '@cowprotocol/common-utils'
-import { ButtonPrimary, CenteredDots } from '@cowprotocol/ui'
 import { useWalletInfo } from '@cowprotocol/wallet'
 import { Currency } from '@uniswap/sdk-core'
-
-import styled from 'styled-components/macro'
 
 import { useErrorModal } from 'legacy/hooks/useErrorMessageAndModal'
 import { useToggleWalletModal } from 'legacy/state/application/hooks'
@@ -21,18 +17,8 @@ import {
 
 import { CurrencySelectButton } from 'common/pure/CurrencySelectButton'
 
-import { ConnectWalletButton, NoFunds, Wrapper } from './styled'
-
-import { useFetchTokenBalance } from '../../hooks/useFetchTokenBalance'
-import { useRecoverFundsCallback } from '../../hooks/useRecoverFundsCallback'
-import { RecoverSigningStep, useRecoverFundsFromProxy } from '../../hooks/useRecoverFundsFromProxy'
-import { BalanceToRecover } from '../../pure/BalanceToRecover'
-
-const ButtonPrimaryStyled = styled(ButtonPrimary)`
-  &:disabled {
-    font-size: 14px;
-  }
-`
+import { RecoverFundsButtons } from './RecoverFundsButtons'
+import { ConnectWalletButton, Wrapper } from './styled'
 
 const selectedCurrencyAtom = atom<Currency | undefined>(undefined)
 
@@ -49,24 +35,13 @@ export function RecoverFundsWidget({ defaultToken: defaultTokenToRefund }: Recov
   const { ErrorModal } = useErrorModal()
   const { open: isSelectTokenWidgetOpen } = useSelectTokenWidgetState()
 
-  const selectedTokenAddress = selectedCurrency ? getCurrencyAddress(selectedCurrency) : undefined
-  const isNativeToken = !!selectedCurrency && getIsNativeToken(selectedCurrency)
-
   const onSelectToken = useOpenTokenSelectWidget()
 
   const sourceChainId = useSourceChainId()
 
-  const { isLoading: isBalanceLoading, tokenBalance } = useFetchTokenBalance(selectedCurrency, sourceChainId)
-  const recoverFundsContext = useRecoverFundsFromProxy(selectedTokenAddress, tokenBalance, isNativeToken)
-  const { txSigningStep } = recoverFundsContext
-
-  const recoverFunds = useRecoverFundsCallback(recoverFundsContext)
-
   const onCurrencySelectClick = useCallback(() => {
     onSelectToken(selectedCurrency, undefined, undefined, setSelectedCurrency)
   }, [onSelectToken, selectedCurrency, setSelectedCurrency])
-
-  const hasBalance = !!tokenBalance?.greaterThan(0)
 
   return (
     <Wrapper>
@@ -85,25 +60,7 @@ export function RecoverFundsWidget({ defaultToken: defaultTokenToRefund }: Recov
             </div>
           )}
 
-          {selectedTokenAddress && (
-            <>
-              <BalanceToRecover tokenBalance={tokenBalance} isBalanceLoading={isBalanceLoading} />
-              {txSigningStep || hasBalance ? (
-                <ButtonPrimaryStyled onClick={recoverFunds} disabled={!hasBalance || !!txSigningStep}>
-                  {txSigningStep && (
-                    <>
-                      {txSigningStep === RecoverSigningStep.SIGN_RECOVER_FUNDS && '1/2 Confirm funds recovering'}
-                      {txSigningStep === RecoverSigningStep.SIGN_TRANSACTION && '2/2 Sign transaction'}
-                      <CenteredDots smaller />
-                    </>
-                  )}
-                  {!txSigningStep && 'Recover funds'}
-                </ButtonPrimaryStyled>
-              ) : (
-                <NoFunds>No funds to recover</NoFunds>
-              )}
-            </>
-          )}
+          <RecoverFundsButtons selectedCurrency={selectedCurrency} sourceChainId={sourceChainId} />
         </>
       )}
     </Wrapper>
